fixed createUserHandler validator + small refactoring
