+++ assets/lambda/src/helpers/event-bridge/system-event-bridge-repository.ts
@@ -1,10 +1,10 @@
 import { EventBridge } from 'aws-sdk';
-import { getEnvironmentSettingsKey } from '../validation-helpers';
 import { UserLambdaSettings } from '../../../../../cdk/settings/lambda-settings';
 import { eventBridgeClient } from './event-bridge-client';
 import { PutEventsRequest } from 'aws-sdk/clients/eventbridge';
 import * as log from 'lambda-log';
 import { SystemEventStorePushEvent } from '../../events/system-event-store-push-event';
+import { getEnvironmentSettingsKey } from '../get-environment-settings-key';
 
 export class SystemEventBridgeRepository {
     private eventBridge: EventBridge;
